% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maeFormula.R
\name{maeFormula}
\alias{maeFormula}
\title{Get the mean absolute error}
\usage{
maeFormula(predicted_list, actual_list)
}
\arguments{
\item{predicted_list}{the predicted results stored in a list without NA/NULL values. The reason why this parameter called predicted_list is that this is the list for predicted results.}

\item{actual_list}{the actual values stored in a list without NA/NULL values. The reason why this parameter called actual_list is that this is the list for actual values.}
}
\value{
mae  the computed mean absolute error result if all inputs are correct.
}
\description{
The function maeFormula is to calculate the mean absolute error or MAE based on the predicted results and their actual values you give. It returns MAE if all inputs have the same length, no NA/NULL values and the type of both lists is numeric, otherwise the function will not work and return an error message.
}
\examples{
predicted_list_1 = list(1,2,3)
actual_list_1 = list(4,5,6)
maeFormula(predicted_list_1, actual_list_1)
}
